seperated mapStateToProps and mapDispatchToProps to Root.js
