+++ src/containers/App.js
@@ -1,10 +1,9 @@
 import React, { useEffect } from "react";
-import { getTodos, setSearchText, setTodosList } from "./actions";
-import { connect } from "react-redux";
-import "./App.css";
+import "./styles/App.css";
+
 import { DataTable, Paginator, SearchField } from "lucid-ui";
-import { actionTypes } from "./actions/actionTypes";
-import { columns } from "./utils/columns";
+import { columns } from "../utils/columns";
+
 const {
   EmptyStateWrapper,
   EmptyStateWrapper: { Title },
@@ -105,29 +104,4 @@ const App = ({
   );
 };
 
-const mapStateToProps = (store) => {
-  const {
-    todos,
-    page,
-    page_size,
-    todos_list,
-    error,
-    search_text,
-  } = store.reducer;
-  return { todos, page, page_size, todos_list, error, search_text };
-};
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    fetchTodos: () => dispatch(getTodos()),
-    setPage: (page) => dispatch({ type: actionTypes.SET_PAGE, payload: page }),
-    setPageSize: (pageSize) =>
-      dispatch({ type: actionTypes.SET_PAGE_SIZE, payload: pageSize }),
-    setTodosList: (todos, page, pageSize) =>
-      dispatch(setTodosList(todos, page, pageSize)),
-    setSearch: (value) =>
-      dispatch({ type: actionTypes.SET_SEARCH_TEXT, payload: value }),
-    fetchSearchedTodos: (text) => dispatch(setSearchText(text)),
-  };
-};
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default App;
